split 4bands page
