+++ src/color-code.gresource.xml
@@ -2,6 +2,7 @@
 <gresources>
   <gresource prefix="/com/oyajun/ColorCode">
     <file preprocess="xml-stripblanks">window.ui</file>
+    <file preprocess="xml-stripblanks">page_4_bands.ui</file>
     <file preprocess="xml-stripblanks">gtk/help-overlay.ui</file>
   </gresource>
 </gresources>
