+++ src/meson.build
@@ -6,6 +6,7 @@ blueprints = custom_target('blueprints',
   input: files(
     'window.blp',
     'gtk/help-overlay.blp',
+    'page_4_bands.blp'
   ),
   output: '.',
   command: [find_program('blueprint-compiler'), 'batch-compile', '@OUTPUT@', '@CURRENT_SOURCE_DIR@', '@INPUT@'],
@@ -41,6 +42,7 @@ color_code_sources = [
   'main.py',
   'window.py',
   'utils.py',
+  'page_4_bands.py'
 ]
 
 install_data(color_code_sources, install_dir: moduledir)
