+++ src/page_4_bands.blp
@@ -1,0 +1,37 @@
+using Gtk 4.0;
+
+template $Page4Bands : Gtk.Box {
+  orientation: vertical;
+  halign: center;
+  valign: center;
+  spacing: 24;
+  margin-top: 12;
+  margin-bottom: 12;
+  margin-start: 12;
+  margin-end: 12;
+
+  CenterBox {
+    halign: center;
+
+    center-widget: Label result_label {
+      width-request: 200;
+      styles [ "title-1" ]
+    };
+
+    end-widget: Button copy_button {
+      icon-name: "edit-copy-symbolic";
+      tooltip-text: _("Copy");
+      valign: center;
+    };
+  }
+
+  Box drop_down_box {
+    halign: center;
+    spacing: 10;
+
+    DropDown drop_down_1 { width-request: 120; }
+    DropDown drop_down_2 { width-request: 120; }
+    DropDown drop_down_3 { width-request: 120; }
+    DropDown drop_down_4 { width-request: 120; }
+  }
+}
