+++ src/window.blp
@@ -14,7 +14,7 @@ template $ColorCodeWindow : Adw.ApplicationWindow {
     setters {
       header_bar.title-widget: null;
       switcher_bar.reveal: true;
-      drop_down_box.orientation: vertical;
+      //drop_down_box.orientation: vertical;
     }
   }
 
@@ -44,54 +44,7 @@ template $ColorCodeWindow : Adw.ApplicationWindow {
         icon-name: "edit-copy-symbolic";
         use-underline: true;
 
-        child: Box {
-          orientation: vertical;
-          halign: center;
-          valign: center;
-          spacing: 24;
-          margin-top: 12;
-          margin-bottom: 12;
-          margin-start: 12;
-          margin-end: 12;
-
-          CenterBox {
-            halign: center;
-
-            center-widget: Label result_label {
-              width-request: 200;
-              styles [
-                "title-1"
-              ]
-            };
-
-            end-widget: Button copy_button {
-              icon-name: "edit-copy-symbolic";
-              tooltip-text: _("Copy");
-              valign: center;
-            };
-          }
-
-          Box drop_down_box {
-            halign: center;
-            spacing: 10;
-
-            DropDown drop_down_1 {
-              width-request: 120;
-            }
-
-            DropDown drop_down_2 {
-              width-request: 120;
-            }
-
-            DropDown drop_down_3 {
-              width-request: 120;
-            }
-
-            DropDown drop_down_4 {
-              width-request: 120;
-            }
-          }
-        };
+        child: $Page4Bands {};
       }
 
       Adw.ViewStackPage page2 {
